Tooltip: Move to Internal Tooltip
This reverts commit 1a2d440f05fde2f3c821d45537a4125eca6696ba.
